inference/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inference"
version = "0.1.0"
description = "SigLIP2 inference service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.118.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "torch>=2.9.1",
  "torchvision>=0.24.1",
  "transformers>=4.57.6",
  "pillow>=10.0.0",
  "python-multipart>=0.0.21"
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
